+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to this project will be documented in this file.
 
+# [0.10.4] - 20-03-2019
+
+### Fixes
+
+- Make Input ref prop optional in Input component
+
+# [0.10.3] - 20-03-2019
+
+### Fixes
+
+- Make calendarInfo and isOutsideRange DatePicker props optional
+- Change the types of startDate and endDate (from string to Moment)
+- Forward ref to Input component
+
 # [0.10.2] - 19-03-2019
 
 ### Changes